+++ quiz_quest/lib/app/features/quiz_pages/films_quiz_pages/hard_films_quiz_page/hard_films_question_widget.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:quiz_quest/app/data/data_sources/quiz_data_source/quiz_categories_data_source.dart';
+import 'package:quiz_quest/app/data/data_sources/user_data_source/user_data_source.dart';
+import 'package:quiz_quest/app/domain/models/films_model/films_quiz_model.dart';
+import 'package:quiz_quest/app/domain/repositories/quiz_repository/quiz_repository.dart';
+import 'package:quiz_quest/app/domain/repositories/user_repository/user_repository.dart';
+import 'package:quiz_quest/app/features/quiz_pages/films_quiz_pages/cubit/films_cubit.dart';
+import 'package:quiz_quest/app/features/quiz_pages/films_quiz_pages/hard_films_quiz_page/hard_lost_life_page.dart';
+import 'package:quiz_quest/app/features/quiz_pages/films_quiz_pages/hard_films_quiz_page/resume_hard_question_quiz_page.dart';
+import 'package:quiz_quest/app/features/quiz_pages/films_quiz_pages/medium_films_quiz_page/medium_lost_life_page.dart';
+import 'package:quiz_quest/app/features/quiz_pages/films_quiz_pages/medium_films_quiz_page/resume_medium_question_quiz_page.dart';
+
+import 'package:quiz_quest/app/features/quiz_pages/quiz_countdown_timer/quiz_countdown_timer.dart';
+import 'package:quiz_quest/app/injection_container.dart';
+
+
+class HardFilmsQuestionWidget extends StatelessWidget {
+  const HardFilmsQuestionWidget({
+    required this.question,
+    super.key,
+  });
+
+  final String question;
+
+  @override
+  Widget build(BuildContext context) {
+    final modifiedQuestion = question
+        .replaceAll('&quot;', '')
+        .replaceAll('&#039;', '')
+        .replaceAll('&aacute;', '')
+        .replaceAll('&ntilde;', '')
+        .replaceAll('&amp;', '')
+        .replaceAll('&rsquo;', '')
+        .replaceAll('oacute;n', '')
+        .replaceAll('&egrave;', '');
+
+    return Center(
+      child: Text(
+        modifiedQuestion,
+        style: GoogleFonts.aBeeZee(
+            fontSize: 24, color: Colors.white, fontWeight: FontWeight.bold),
+        textAlign: TextAlign.center,
+      ),
+    );
+  }
+}
